Added ability to edit bio for users. Added ability to see other users profiles.
Pair Programmed with Aaron.

+++ Hash/app/Http/Controllers/User.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Entities;
+use App\Repositories\Users;
 use Doctrine\ORM\EntityManagerInterface;
 use Illuminate\Contracts\Filesystem\Filesystem;
 use Illuminate\Routing\Controller;
@@ -12,22 +13,39 @@ use Illuminate\Contracts\Auth\Guard;
 
 class User extends Controller
 {
-	public function view(Guard $auth, EntityManagerInterface $em, $user = null)
+	public function view(Request $request, Guard $auth, Users $users, $username = null)
 	{
-		if ($user) {
-			dd($user);
-		} else if ($auth->check()) {
-			$user = $auth->user();
+		$user = null;
 
+		if ($username)
+		{
+			$user = $users->findOneByUsername($username);
+		}
+		else if ($auth->check())
+		{
+			$user = $auth->user();
+		}
+		else
+		{
+			$request->session()->flash("alert-danger", trans('auth.required'));
+			return redirect("/auth/login");
+		}
+
+		if($user instanceof Entities\User)
+		{
 			return view('user.view')->with(compact('user'));
-		} else {
-			dd('Not authenticated, no user.');
+		}
+		else
+		{
+			abort(404);
 		}
 	}
 
-	public function form()
+	public function form(Guard $auth)
 	{
-		return view('user.form');
+		/** @var Entities\User $user */
+		$user = $auth->user();
+		return view('user.form')->with(compact('user'));
 	}
 
 	/**
@@ -45,17 +63,23 @@ class User extends Controller
 		$user = $auth->user();
 
 		$valid = $validator->make($request->all(), [
-			'avatar' => "required|image|dimensions:max_width=800,max_height=800",
+			'avatar' => "image|dimensions:max_width=800,max_height=800",
+			'bio' => "max:300"
 		]);
 
 		$valid->validate();
+		$data = $valid->getData();
 
-		if ($user->getPicture())
-		{
-			$file->delete($user->getPicture());
+		if($request->exists("avatar")) {
+			if ($user->getPicture()) {
+				$file->delete($user->getPicture());
+			}
+
+			$user->setPicture($request->file('avatar')->storePublicly("public/avatars"));
 		}
 
-		$user->setPicture($request->file('avatar')->storePublicly("public/avatars"));
+		$user->setBio($data['bio']);
+
 		$em->flush();
 
 		return redirect('/user');
